feat: convert display_order to BigDecimal, add upgrade-track and scale test coverage
Closes out the remaining open items from research/Kimball_type_2.md's
Implementation Readiness Notes:

- Convert Section.displayOrder, Step.displayOrder, StepsSections.step/
  sectionDisplayOrder, SectionsQuestion.displayOrder, and Answer.stepId/
  sectionId from Integer to BigDecimal, matching the NUMERIC columns the
  migration already created. Fixed the call sites that feed these into
  DisplayKey's int-based setters (.intValue() -- its fixed-width zero-
  padded string encoding doesn't support decimals yet, a future Author
  Tool concern) and Answer's two write sites.

- Add ManualSchemaMigratorUpgradeTest: seeds a pre-V010 "v2.x" history
  via db/migration-v3's V001-V009, confirms it fails validation against
  db/migration (checksum mismatch), runs the real upgrade path, repairs,
  and confirms a second boot validates cleanly. Previously the only path
  ManualSchemaMigrator has -- the ALTER-based upgrade branch -- had zero
  automated coverage; every existing test only ever exercises the
  greenfield track. Needed adding flyway-database-postgresql as a test
  dependency since a plain-JUnit test calling the Flyway API directly
  doesn't get Quarkus's bundled Postgres support.

- Add ManualSchemaMigratorScaleTest (@Disabled by default -- seeds ~300k
  rows): times the V010 upgrade migration against a synthetic dataset
  sized so answers dominates, addressing the doc's "test against a
  production-sized snapshot" note. Measured ~2.1s locally; documented as
  a synthetic proxy pending a real anonymized-snapshot run before
  go-live.

- Add docs/plans/kimball-stakeholder-review.md: a plain-language
  decision summary (not sign-off itself) of the two clinically-relevant
  behaviors the doc flags as needing research/compliance sign-off --
  retroactive ontology reclassification and SCD Type 1 relabeling of
  dim_step/dim_section.

Full suite: 342 tests, 0 failures.

+++ src/main/java/com/elicitsoftware/QuestionManager.java
@@ -512,7 +512,7 @@ public class QuestionManager {
             for (SectionsQuestion sectionsQuestion : sectionsQuestions) {
                 // Build a display key for the answer
                 dkey = new DisplayKey(key.getValue());
-                dkey.setQuestion(sectionsQuestion.displayOrder);
+                dkey.setQuestion(sectionsQuestion.displayOrder.intValue());
                 answer = new Answer(dkey, sectionsQuestion, sectionsQuestion.question.text, respondentId, sectionsQuestion.question.defaultValue);
                 answer = saveAnswer(answer, null);
                 if ("HTML".equals(sectionsQuestion.question.questionType.name) || answer.getTextValue() != null) {
@@ -609,7 +609,7 @@ public class QuestionManager {
                             if (relationship.downstreamQuestion != null) {
                                 DisplayKey key = buildDisplayKey(upstreamAnswer, relationship);
 
-                                key.setQuestion(relationship.downstreamQuestion.displayOrder);
+                                key.setQuestion(relationship.downstreamQuestion.displayOrder.intValue());
                                 // this may be the first answer in a section
                                 buildSectionAnswer(upstreamAnswer.respondentId, key.getSectionString(), dependents);
                                 Answer a = new Answer(key, relationship.downstreamQuestion,
@@ -670,7 +670,7 @@ public class QuestionManager {
                     if (relationship.downstreamQuestion != null) {
                         // a question
                         DisplayKey key = buildDisplayKey(upstreamAnswer, relationship);
-                        key.setQuestion(relationship.downstreamQuestion.displayOrder);
+                        key.setQuestion(relationship.downstreamQuestion.displayOrder.intValue());
                         Answer a = getAnswerByDisplayKey(upstreamAnswer.respondentId, key.getValue(), false);
                         if (a != null) {
                             dependents.put(relationship.id,
@@ -698,7 +698,7 @@ public class QuestionManager {
                     if (relationship.downstreamQuestion != null) {
                         // a question
                         DisplayKey key = buildDisplayKey(upstreamAnswer, relationship);
-                        key.setQuestion(relationship.downstreamQuestion.displayOrder);
+                        key.setQuestion(relationship.downstreamQuestion.displayOrder.intValue());
                         Answer a = getAnswerByDisplayKey(upstreamAnswer.respondentId, key.getValue(), false);
                         if (a != null) {
                             saveAnswer(a, dependents);
@@ -1022,7 +1022,7 @@ public class QuestionManager {
             Answer answer;
             for (SectionsQuestion sectionsQuestion : initial) {
                 DisplayKey dkey = new DisplayKey(sectionKey.getValue());
-                dkey.setQuestion(sectionsQuestion.displayOrder);
+                dkey.setQuestion(sectionsQuestion.displayOrder.intValue());
                 answer = new Answer(dkey, sectionsQuestion, sectionsQuestion.question.text,
                         upstreamAnswer.respondentId, sectionsQuestion.question.defaultValue);
                 answer = saveAnswer(answer, dependents);
@@ -1053,7 +1053,7 @@ public class QuestionManager {
         DisplayKey answerKey = new DisplayKey(upstreamAnswer.getDisplayKey());
         answerKey.setStep(getStepDisplayOrder(relationship.surveyId, relationship.downstreamStep.id, respondentId));
         answerKey.setSection(getSectionDisplayOrder(relationship.surveyId, relationship.downstreamSection.id, respondentId));
-        answerKey.setQuestion(relationship.downstreamQuestion.displayOrder);
+        answerKey.setQuestion(relationship.downstreamQuestion.displayOrder.intValue());
 
         List<Answer> answers = Answer.findByAnswerQueryString(respondentId, answerKey.getAnswerQueryString());
 
@@ -1072,7 +1072,7 @@ public class QuestionManager {
                 key.setSection(getSectionDisplayOrder(relationship.surveyId, relationship.downstreamSection.id, respondentId));
             }
             if (downstreamQuestion.displayOrder != null) {
-                key.setQuestion(downstreamQuestion.displayOrder);
+                key.setQuestion(downstreamQuestion.displayOrder.intValue());
             }
             DisplayKey newKey;
             for (int i = answers.size() + 1; i <= repeatValue; i++) {
@@ -1589,7 +1589,7 @@ public class QuestionManager {
             // Section
             int stepId;
             if (relationship.upstreamStep == null) {
-                stepId = answer.stepId;
+                stepId = answer.stepId.intValue();
             } else {
                 stepId = relationship.upstreamStep.id;
             }
@@ -1606,7 +1606,7 @@ public class QuestionManager {
             // Step
             int stepId;
             if (relationship.upstreamStep == null) {
-                stepId = answer.stepId;
+                stepId = answer.stepId.intValue();
             } else {
                 stepId = relationship.upstreamStep.id;
             }
